Update blockchain with http test
* add command to test /transaction
* add command to test /mine

+++ blockchain/transaction.go
@@ -5,9 +5,9 @@ import (
 )
 
 type Transaction struct {
-	Sender   string `json:"from" binding:"required"`
-	Receiver string `json:"to" binding:"required"`
-	Amount   uint32 `json:"amount" binding:"required"`
+	Sender   string  `json:"from" binding:"required"`
+	Receiver string  `json:"to" binding:"required"`
+	Amount   float32 `json:"amount" binding:"required"`
 }
 
 func (t *Transaction) Inspect() {
